+++ src/main/java/com/gobra/sistemagestaodeobras/dto/SocioResponseDTO.java
@@ -15,7 +15,7 @@ public record SocioResponseDTO (
 
   Date dataDesativacao,
 
-  Integer telegramID,
+  Long telegramID,
 
   List<Compra> compras,
